refactor: UI & Service 분리 (#107)
