+++ src/features/index.ts
@@ -1,7 +1,8 @@
 export * from './DashBoardTemplate'
-export * from './home/HelperBox'
-export * from './home/InfoBox'
-export * from './home/SectionWrapper'
+export * from './home/ui/HelperBox'
+export * from './home/ui/HomePage'
+export * from './home/ui/InfoBox'
+export * from './home/ui/SectionWrapper'
 export * from './sign-in/SignInStep'
 export * from './sign-up/AccountStep'
 export * from './sign-up/CompleteStep'
