+++ src/hooks/apis/useHome.ts
@@ -1,11 +1,11 @@
 import { useQuery } from '@tanstack/react-query'
 
 import { getHomePageData } from '@/apis'
-import type { HomeParamsType, HomeResponseType } from '@/types'
+import type { HomeResponseType } from '@/types'
 
-export const useHome = ({ userId, localDate }: HomeParamsType) => {
+export const useHome = (userId: string) => {
   return useQuery<HomeResponseType>({
     queryKey: [userId],
-    queryFn: () => getHomePageData({ userId, localDate }),
+    queryFn: () => getHomePageData(userId),
   })
 }
